requirements.txt
click==6.7
Django==1.11.2
django-bootstrap3==8.2.3
django-fontawesome==0.3.1
first==2.0.1
olefile==0.44
Pillow==4.1.1
pytz==2017.2
PyYAML==3.12
six==1.10.0
